Refactor encode tags iterator for reuse
